requirements.txt
asgiref==3.2.3
attrs==19.3.0
Django==3.0
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-stubs==1.0.0
importlib-metadata==1.2.0
jsonschema==3.2.0
more-itertools==8.0.0
pyrsistent==0.15.6
pytz==2019.3
PyYAML==5.2
six==1.13.0
sqlparse==0.3.0
zipp==0.6.0
